Speed
Add moving with speed, speed transforms,  inverse and det functions for every size of matrix.
Add some printing and key checking functions.
The problem is that when the robot is in the last available point of trajectory determinant of Jacobian is not a zero and the speed and movements are strange